Update default data
Bug fix:
• Temporarily removed two datasets that are not correctly loading from the Klepikova data until the dataset link loads correctly
• Added another level of verification before deleting a user

In progress:
• MEDIAN control data
• Getting the eFP-Seq Browser to work for internet explorer

TODO:
• Make eFP Overview more responsive to the screen's resolution
• TMM or TPM and median for additional normalization methods
• Expand the creation of the private database to include adding from public and private database to your creation

Known bugs:
• Does not work on IE
• Google Drive repos do not work for IGB links
• Download page as an image does not render SVGs on Microsoft Edge

+++ cgi-bin/core/custom.js
@@ -1979,7 +1979,7 @@ function delete_selectedXML() {
  * Confirm the action of delete all users and if so, begin deletion process
  */
 function confirm_deleteUser() {
-  if (users_email != "" && users_email === gapi.auth2.getAuthInstance().currentUser.Ab.w3.U3) {
+  if (users_email != "" && users_email === gapi.auth2.getAuthInstance().currentUser.Ab.w3.U3 && $('#logoutModal').is(':visible')) {
     delete_allXMLs(gapi.auth2.getAuthInstance().currentUser.Ab.w3.Eea);
   }
 }
